backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
celery[redis]
redis
langchain
langgraph
python-dotenv
python-multipart
docker
pandas
groq
matplotlib
seaborn
python-jose
passlib[bcrypt]